- fixes message codes - adds lacking entries in jdiam cca sessions - adds messsage to cca handlers/listeners (STR/STA,ACR/ACA,ASR/ASA + extension)
git-svn-id: https://mobicents.googlecode.com/svn/trunk/servers/diameter@3994 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ core/jdiameter/impl/src/main/java/org/jdiameter/common/api/app/cca/IClientCCASessionContext.java
@@ -10,8 +10,8 @@ public interface IClientCCASessionContext {
 
 	void txTimerExpired(ClientCCASession session);
 	
-	long getDefaultCCFHValue();
-	long getDefaultDDFHValue();
+	int getDefaultCCFHValue();
+	int getDefaultDDFHValue();
 
 	void grantAccessOnDeliverFailure(ClientCCASession clientCCASessionImpl,Message request);
